Extract ListInterfaceControllerViewModel
